Move VariablySizedColumns to executor crate
