+++ backend/src/estark/mod.rs
@@ -15,8 +15,11 @@ use std::{
 use crate::{Backend, BackendFactory, BackendOptions, Error, Proof};
 use powdr_ast::analyzed::Analyzed;
 
-use powdr_executor::witgen::WitgenCallback;
-use powdr_number::{get_only_size_cloned, DegreeType, FieldElement, VariablySizedColumn};
+use powdr_executor::{
+    constant_evaluator::{get_only_size_cloned, VariablySizedColumn},
+    witgen::WitgenCallback,
+};
+use powdr_number::{DegreeType, FieldElement};
 use serde::Serialize;
 use starky::types::{StarkStruct, Step, PIL};
