+++ backend/src/estark/starky_wrapper.rs
@@ -5,10 +5,11 @@ use std::time::Instant;
 
 use crate::{Backend, BackendFactory, BackendOptions, Error};
 use powdr_ast::analyzed::Analyzed;
-use powdr_executor::witgen::WitgenCallback;
-use powdr_number::{
-    get_only_size_cloned, FieldElement, GoldilocksField, LargeInt, VariablySizedColumn,
+use powdr_executor::{
+    constant_evaluator::{get_only_size_cloned, VariablySizedColumn},
+    witgen::WitgenCallback,
 };
+use powdr_number::{FieldElement, GoldilocksField, LargeInt};
 
 use starky::{
     merklehash::MerkleTreeGL,
