+++ executor/Cargo.toml
@@ -20,6 +20,7 @@ bit-vec = "0.6.3"
 num-traits = "0.2.15"
 lazy_static = "1.4.0"
 indicatif = "0.17.7"
+serde = { version = "1.0", default-features = false, features = ["alloc", "derive", "rc"] }
 
 [dev-dependencies]
 test-log = "0.2.12"
