+++ number/src/lib.rs
@@ -8,10 +8,6 @@ mod bn254;
 mod goldilocks;
 mod serialize;
 mod traits;
-
-use std::collections::BTreeMap;
-
-use serde::{Deserialize, Serialize};
 pub use serialize::{
     buffered_write_file, read_polys_csv_file, write_polys_csv_file, CsvRenderMode, ReadWrite,
 };
@@ -31,52 +27,6 @@ pub type DegreeType = u64;
 pub fn log2_exact(n: BigUint) -> Option<usize> {
     n.trailing_zeros()
         .filter(|zeros| n == (BigUint::from(1u32) << zeros))
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct VariablySizedColumn<F> {
-    column_by_size: BTreeMap<usize, Vec<F>>,
-}
-
-#[derive(Debug)]
-pub struct HasMultipleSizesError;
-
-impl<F> VariablySizedColumn<F> {
-    /// Create a view where each column has a single size. Fails if any column has multiple sizes.
-    pub fn get_only_size(&self) -> Result<&Vec<F>, HasMultipleSizesError> {
-        if self.column_by_size.len() != 1 {
-            return Err(HasMultipleSizesError);
-        }
-        Ok(self.column_by_size.values().next().unwrap())
-    }
-}
-
-pub fn get_only_size<F>(
-    column: &[(String, VariablySizedColumn<F>)],
-) -> Result<Vec<(String, &Vec<F>)>, HasMultipleSizesError> {
-    column
-        .iter()
-        .map(|(name, column)| Ok((name.clone(), column.get_only_size()?)))
-        .collect()
-}
-
-pub fn get_only_size_cloned<F: Clone>(
-    column: &[(String, VariablySizedColumn<F>)],
-) -> Result<Vec<(String, Vec<F>)>, HasMultipleSizesError> {
-    get_only_size(column).map(|column| {
-        column
-            .into_iter()
-            .map(|(name, column)| (name, column.clone()))
-            .collect()
-    })
-}
-
-impl<F> From<Vec<F>> for VariablySizedColumn<F> {
-    fn from(column: Vec<F>) -> Self {
-        VariablySizedColumn {
-            column_by_size: [(column.len(), column)].into_iter().collect(),
-        }
-    }
 }
 
 #[cfg(test)]
